feat: 카드번호 훅에 카드번호 상태 추가
